Extract BitchatMessage into a separate file
